+++ packages/react-native-eyecandy-icons/src/icons/ChevronDown.tsx
@@ -1,0 +1,19 @@
+import React, { memo } from 'react';
+
+import { Path } from 'react-native-svg';
+
+import { SvgIconContainer } from '../core';
+import { SvgIconProps } from '../types';
+
+const Icon = (props: SvgIconProps) => (
+  <SvgIconContainer {...props}>
+    <Path
+      d="m6 9 6 6 6-6"
+      strokeWidth={1.5}
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </SvgIconContainer>
+);
+
+export const ChevronDown = memo(Icon);