+++ packages/react-native-eyecandy-icons/src/icons/Dots.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+
+import { Path } from 'react-native-svg';
+
+import { SvgIconContainer } from '../core';
+import { SvgIconProps } from '../types';
+
+const Icon = (props: SvgIconProps) => (
+  <SvgIconContainer {...props}>
+    <Path
+      d="M5 13a1 1 0 1 0 0-2 1 1 0 0 0 0 2ZM12 13a1 1 0 1 0 0-2 1 1 0 0 0 0 2ZM19 13a1 1 0 1 0 0-2 1 1 0 0 0 0 2Z"
+      strokeWidth={2}
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </SvgIconContainer>
+);
+
+export const Dots = React.memo(Icon);