+++ packages/react-native-eyecandy-icons/src/icons/MoonSun.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import { Path } from 'react-native-svg';
+
+import { SvgIconContainer } from '../core';
+import { SvgIconProps } from '../types';
+
+const Icon = (props: SvgIconProps) => (
+  <SvgIconContainer variant="fill" {...props}>
+    <Path
+      d="M12 22c5.523 0 10-4.477 10-10S17.523 2 12 2 2 6.477 2 12s4.477 10 10 10Zm0-2V4a8 8 0 0 1 0 16Z"
+      fill="currentColor"
+    />
+  </SvgIconContainer>
+);
+
+export const MoonSun = React.memo(Icon);