+++ packages/react-native-eyecandy-icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nomada-sh/react-native-eyecandy-icons
 
+## 0.0.9
+
+### Patch Changes
+
+- cd0e21f: Removed unnecessary props from IconProps.
+
 ## 0.0.8
 
 ### Patch Changes